:recycle: rename many variables and optimize imports
